+++ My_Study_New/My_Study/Business/Module/DebugLog/controller/MDLogViewController.m
@@ -245,7 +245,7 @@
         case FromPageTypeDeault:
         {
             *titleName = logModel.context;
-            *subTitleName = [NSString stringWithFormat:@"%ld",logModel.count];
+            *subTitleName = [NSString stringWithFormat:@"%lu",(unsigned long)logModel.count];
         }
             break;
         case FromPageTypeOne:
